Debian 8 (jessie) became obsolete stable release
In June 2018 Debian maintainers stop supporting jessie release.
As part of this process, all jessie-update packages have been
merged and, since March 2019, the repositories have been moved
to archive.debian.org servers.
This means that line 5 and 6 of sources.list file are no longer
required and the base repository address must be updated.
Also, due to the new repository servers management, the Release
files are no longer updated and an additional hack is required to
ignore the validity check.