var_make_debian: add common framework for fetching git repositories
Currently, git repos are defined in the machine files, and then
var_make_debian checks if each repository is defined and then fetches
them.

Let's consolidate this by adding a list of repositories to fetch in
each machine file, and then iterate through the list in var_make_debian.sh

This adds flexibility for working with different SOC_FAMILYs, simplifies
the cmd_make_deploy function, and consolidates the code to fetch
and patch a git repository.

Signed-off-by: Nate Drude <nate.d@variscite.com>