var_make_debian: add simple state machine for resuming make rootfs
The 'rootfs' make target takes a long time to build and if interrupted,
must start over from the beginning.

This patch breaks the make_debian_weston_rootfs function into many smaller
steps. When a step is completed, it is written to .build_steps. The
build can then be resumed using the following logic:

1. Before running a step, check .build_steps for the step.
2. If it's already complete, skip it
3. If it's not complete, run the step and write the step to .build_steps
when complete.

Notice:

The main goal of this is to improve the development workflow by allowing
builds to resume if a step fails. Once a step is complete, it will not run
again until 'make clean'. It does not check if a steps source code or
files dependencies have changed.

Production images should be built from a clean working directory.

Signed-off-by: Nate Drude <nate.drude@variscite.com>