weston_rootfs: make_weston_sdcard_am6: adjust PART parameter to make sdimage
Since the patch:
  https://github.com/varigit/meta-variscite-sdk-ti/commit/02214ded28786190552372c2254fd4d06ce1ea90
am6_install_yocto.sh is used to make sdcard and sdimage when ran from
x86 host machine and also is used to install yocto/debian images on eMMC
when ran from target.

This change fixes the issue when the script is used to make a recovery-sd:
  sudo MACHINE=am62x-var-som ./var_make_debian.sh -c sdimage -i <sdimage_name>
where the loop device is used.

Signed-off-by: FrancescoFerraro <francesco.f@variscite.com>