VAR-SOM-AM62: Debian: update loop device management
When calling the function make_weston_sdcard_am6 using the 'sdimage'
command, the 'p' partition prefix is automatically used.
However, the function can be also called when a user creates a custom
loop device and use the 'sdcard' command.
In this 2nd condition, the 'p' partition prefix must be added by
checking the block name, as done in make_weston_sdcard_imx.
The command check is not really needed.

Signed-off-by: Pierluigi Passaro <pierluigi.p@variscite.com>